refactor(sonar): deduplicate operation-pipeline test interceptors
Extract the copy-pasted private interceptor helpers (CountingInterceptor,
RecordingInterceptor, MutatingInterceptor, ShortCircuitInterceptor,
ContextCapturingInterceptor, ResultCapturingInterceptor, ItemsPrePost,
ItemsSetting/Reading, CustomResult/CustomOperationResult, SingleKey,
Generic* variants, StaticUserAccessor, TestSystemTime) into a single
shared TestInterceptors.cs, marked internal sealed.

- Fixes S3260 (25x): all private interceptor helpers now sealed
- Fixes S4487 (2x): shared ContextCapturingInterceptor has its
  CapturedTimestamp/CapturedKey read by InterceptorPipelineTests, so
  they are no longer unread
- Fixes S2930 (1x): dispose the CancellationTokenSource in
  EntityOperationContextTests via 'using'
- Suppresses S1694 (1x) on AbstractInterceptor with justification:
  the test deliberately needs an abstract class to verify that
  WithInterceptor rejects non-concrete types
- Drops new-code duplication density from 8.36% to near 0% by removing
  the repeated private helper class declarations across 6 test files
