+++ test/Kista.Manager.XUnit/Unit/Operations/InterceptorPipelineCoverageTests.cs
@@ -703,181 +703,4 @@ public class InterceptorPipelineCoverageTests {
 		Assert.NotNull(interceptor.CapturedOriginal);
 	}
 
-	// --- Test interceptors ---
-
-	private class RecordingInterceptor : IEntityManagerInterceptor<Person, string> {
-		private readonly string _name;
-		private readonly List<string> _callOrder;
-
-		public RecordingInterceptor(string name, List<string> callOrder) {
-			_name = name;
-			_callOrder = callOrder;
-		}
-
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			_callOrder.Add($"{_name}.Pre");
-			return default;
-		}
-
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result) {
-			_callOrder.Add($"{_name}.Post");
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private class MutatingInterceptor : IEntityManagerInterceptor<Person, string> {
-		private readonly Action<Person> _mutate;
-		public MutatingInterceptor(Action<Person> mutate) => _mutate = mutate;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			_mutate(context.Entity);
-			return default;
-		}
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class ShortCircuitInterceptor : IEntityManagerInterceptor<Person, string> {
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context)
-			=> new(OperationResult.Fail(new OperationError("SHORT_CIRCUIT", "Test", "Short-circuited")));
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class CountingInterceptor : IEntityManagerInterceptor<Person, string> {
-		public int PreWriteCount;
-		public int PostWriteCount;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			PreWriteCount++;
-			return default;
-		}
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result) {
-			PostWriteCount++;
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private class ContextCapturingInterceptor : IEntityManagerInterceptor<Person, string> {
-		public string? CapturedActor;
-		public DateTimeOffset CapturedTimestamp;
-		public EntityOperationKind CapturedKind;
-		public Person? CapturedOriginal;
-		public string? CapturedKey;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			CapturedActor = context.Actor;
-			CapturedTimestamp = context.Timestamp;
-			CapturedKind = context.Kind;
-			CapturedOriginal = context.Original;
-			CapturedKey = context.Key;
-			return default;
-		}
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class ResultCapturingInterceptor : IEntityManagerInterceptor<Person, string> {
-		public IOperationResult? CapturedResult;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context)
-			=> default;
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result) {
-			CapturedResult = result;
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private class ItemsPrePostInterceptor : IEntityManagerInterceptor<Person, string> {
-		public bool PostWriteSawPreWriteItem;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			context.Items["pre"] = "set";
-			return default;
-		}
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result) {
-			PostWriteSawPreWriteItem = context.Items.TryGetValue("pre", out var val) && Equals(val, "set");
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private class CustomResultInterceptor : IEntityManagerInterceptor<Person, string> {
-		private readonly IOperationResult _result;
-		public CustomResultInterceptor(IOperationResult result) => _result = result;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context)
-			=> new(_result);
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class SingleKeyCountingInterceptor : IEntityManagerInterceptor<Person> {
-		public int PreWriteCount;
-		public int PostWriteCount;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, object> context) {
-			PreWriteCount++;
-			return default;
-		}
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, object> context, IOperationResult result) {
-			PostWriteCount++;
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private class GenericShortCircuitInterceptor<T, K> : IEntityManagerInterceptor<T, K>
-		where T : class where K : notnull {
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<T, K> context)
-			=> new(OperationResult.Fail(new OperationError("SHORT_CIRCUIT", "Test", "Short-circuited")));
-		public ValueTask PostWriteAsync(IEntityOperationContext<T, K> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class GenericRecordingInterceptor<T, K> : IEntityManagerInterceptor<T, K>
-		where T : class where K : notnull {
-		private readonly string _name;
-		private readonly List<string> _callOrder;
-		public GenericRecordingInterceptor(string name, List<string> callOrder) {
-			_name = name;
-			_callOrder = callOrder;
-		}
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<T, K> context) {
-			_callOrder.Add($"{_name}.Pre");
-			return default;
-		}
-		public ValueTask PostWriteAsync(IEntityOperationContext<T, K> context, IOperationResult result) {
-			_callOrder.Add($"{_name}.Post");
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private class GenericCountingInterceptor<T, K> : IEntityManagerInterceptor<T, K>
-		where T : class where K : notnull {
-		public int PreWriteCount;
-		public int PostWriteCount;
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<T, K> context) {
-			PreWriteCount++;
-			return default;
-		}
-		public ValueTask PostWriteAsync(IEntityOperationContext<T, K> context, IOperationResult result) {
-			PostWriteCount++;
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private sealed class StaticUserAccessor<TKey> : IUserAccessor<TKey> {
-		private readonly TKey? _userId;
-		public StaticUserAccessor(TKey? userId) => _userId = userId;
-		public TKey? GetUserId() => _userId;
-	}
-
-	/// <summary>
-	/// A custom IOperationResult implementation that is not an OperationResult struct,
-	/// to exercise the fallback branches of ToOperationResult.
-	/// </summary>
-	private class CustomOperationResult : IOperationResult {
-		private readonly bool _hasError;
-		private readonly bool _isSuccess;
-
-		public CustomOperationResult(bool hasError, bool isSuccess = false) {
-			_hasError = hasError;
-			_isSuccess = isSuccess;
-		}
-
-		public IOperationError? Error => _hasError ? new OperationError("CUSTOM", "Test", "Custom error") : null;
-		public OperationResultType ResultType => _isSuccess ? OperationResultType.Success : OperationResultType.Error;
-	}
 }
