+++ test/Kista.Manager.XUnit/Unit/Operations/InterceptorPipelineTests.cs
@@ -206,113 +206,4 @@ public class InterceptorPipelineTests {
 		Assert.Equal("key-123", interceptor.CapturedKey);
 	}
 
-	// --- Test interceptors ---
-
-	private class RecordingInterceptor : IEntityManagerInterceptor<Person, string> {
-		private readonly string _name;
-		private readonly List<string> _callOrder;
-		private readonly bool _shortCircuit;
-
-		public RecordingInterceptor(string name, List<string> callOrder, bool shortCircuit = false) {
-			_name = name;
-			_callOrder = callOrder;
-			_shortCircuit = shortCircuit;
-		}
-
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			_callOrder.Add($"{_name}.Pre");
-			if (_shortCircuit)
-				return new ValueTask<IOperationResult?>(OperationResult.Fail(new OperationError("SHORT_CIRCUIT", "Test", "Short-circuited by A")));
-			return default;
-		}
-
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result) {
-			_callOrder.Add($"{_name}.Post");
-			return ValueTask.CompletedTask;
-		}
-	}
-
-	private class MutatingInterceptor : IEntityManagerInterceptor<Person, string> {
-		private readonly Action<Person> _mutate;
-
-		public MutatingInterceptor(Action<Person> mutate) {
-			_mutate = mutate;
-		}
-
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			_mutate(context.Entity);
-			return default;
-		}
-
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class ItemsSettingInterceptor : IEntityManagerInterceptor<Person, string> {
-		private readonly string _key;
-		private readonly object? _value;
-
-		public ItemsSettingInterceptor(string key, object? value) {
-			_key = key;
-			_value = value;
-		}
-
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			context.Items[_key] = _value;
-			return default;
-		}
-
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class ItemsReadingInterceptor : IEntityManagerInterceptor<Person, string> {
-		private readonly string _key;
-		private readonly object? _expectedValue;
-		public bool ReadValueMatched;
-
-		public ItemsReadingInterceptor(string key, object? expectedValue) {
-			_key = key;
-			_expectedValue = expectedValue;
-		}
-
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			ReadValueMatched = context.Items.TryGetValue(_key, out var val) && Equals(val, _expectedValue);
-			return default;
-		}
-
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private class ContextCapturingInterceptor : IEntityManagerInterceptor<Person, string> {
-		public string? CapturedActor;
-		public DateTimeOffset CapturedTimestamp;
-		public EntityOperationKind CapturedKind;
-		public Person? CapturedOriginal;
-		public string? CapturedKey;
-
-		public ValueTask<IOperationResult?> PreWriteAsync(IEntityOperationContext<Person, string> context) {
-			CapturedActor = context.Actor;
-			CapturedTimestamp = context.Timestamp;
-			CapturedKind = context.Kind;
-			CapturedOriginal = context.Original;
-			CapturedKey = context.Key;
-			return default;
-		}
-
-		public ValueTask PostWriteAsync(IEntityOperationContext<Person, string> context, IOperationResult result)
-			=> ValueTask.CompletedTask;
-	}
-
-	private sealed class StaticUserAccessor<TKey> : IUserAccessor<TKey> {
-		private readonly TKey? _userId;
-		public StaticUserAccessor(TKey? userId) => _userId = userId;
-		public TKey? GetUserId() => _userId;
-	}
-
-	private sealed class TestSystemTime : ISystemTime {
-		public DateTimeOffset UtcNow => new DateTimeOffset(2026, 7, 12, 10, 0, 0, TimeSpan.Zero);
-		public DateTimeOffset Now => UtcNow.ToLocalTime();
-	}
 }
